+++ Server/src/Routes/FoodQueue.ts
@@ -1,21 +1,20 @@
 import {Router} from "express";
 import {hasRole, isLogged} from "./Auth";
 import Joi from "joi";
-import DrinkQueue from "../models/DrinkQueue";
 import FoodQueue from "../models/FoodQueue";
 import Orders from "../models/Orders";
-import mongoose from "mongoose";
-import Menus, {IMenu} from "../models/Menus";
 
-
+// Define a schema for food queue input validation using Joi
 export const FoodQueueSchemaValidation = Joi.object().keys({
     dish: Joi.string().required(),
     order: Joi.string().required(),
 })
 
+// Export router
 export default (): Router => {
     const app = Router();
 
+    // GET endpoint to retrieve food in queue based on the query parameters passed
     app.get('/', isLogged, hasRole("Cook", "Cashier"), async (req, res)=> {
         try{
             const foods = await FoodQueue.find(req.query);
@@ -25,6 +24,7 @@ export default (): Router => {
         }
     });
 
+    // GET endpoint to retrieve food in queue by its ID
     app.get('/:id', isLogged, hasRole("Cook", "Cashier"), async (req, res)=> {
         try{
             const food = await FoodQueue.findById(req.params.id).populate("dish");
@@ -35,10 +35,12 @@ export default (): Router => {
         }
     });
 
+    // POST endpoint to modify an existing food in queue
     app.post("/:id", isLogged, hasRole('Cook'), async (req, res) => {
         const food = await FoodQueue.findById(req.params.id);
-        if (food.begin && food.end) return res.status(200).send("Already finished");
+        if(food === null) return res.status(400).send("Food in queue doesn't exist");
 
+        if (food.begin && food.end) return res.status(200).send("Already finished");
 
         if (!food.begin) {
             food.begin = true;
@@ -57,26 +59,27 @@ export default (): Router => {
                 console.log("TUTTO FINITOOOOOOOOOOOOOO")
             }
         }
-        try{
-            food.save()
-        } catch (err) {
-            return res.status(400).send("Something went wrong");
-        }
 
-        return res.status(200).send("Food modified correctly");
+        // Save the changes to the food in the queue in the database
+        try{
+            await food.save()
+            return res.status(200).send("Food modified correctly");
+        } catch (err) {
+            return res.status(400).send(err);
+        }
     });
 
-    //since they will be deleted when the receipt is calculated the role should be cashier right?
+    // DELETE endpoint to delete food in the queue by its ID
     app.delete("/:id", isLogged, hasRole('Cook'), async (req, res) => {
         const food = await FoodQueue.findById(req.params.id);
+        if (food === null) return res.status(400).send("Food in queue doesn't exist");
 
         try{
             await food.deleteOne();
+            return res.status(200).send("Food deleted successfully");
         } catch (err) {
-            return res.status(400).send("Something went wrong");
+            return res.status(400).send(err);
         }
-
-        return res.status(200).send("Food deleted successfully");
     });
 
     return app;
